Fixing Enrolment translateToSOAP and updateFromSOAP implementations.
translateToSOAP wasn't translating all the attributes. It does now. It also now throws an IllegalStateException if the status is not final rather than if null setting it to NEW. i.e., getAllowsReplicationQueueing already has this condition covered so it should never occur.

updateFromSOAP now correctly deals with the incoming data conditionally updating attributes if null and relationships if the record is new only.