[squish 7612]  Fixing shortlist dropping off + _many_ and various other component improvements (e.g., duplicate code moved into shared super classes, awake logic moved to lazy instantiation to resolve race conditions for sub-class loading...)
Also set backtracking page size to 1 (which == off but allows stateful components to continue to function).

Issues to resolve/test:
- what to do when payments fails. [1]
- what to do with session data once payment succeeds [2]
- need to test zero dollar enrolments

[1] Enrolments reserve their places on save with a status == new. The status of the enrolments currently follows that of the payment. So if the payment is marked as failed, so are the enrolments. Is this what we want? 
The choices are:
a. allow the user to try paying again keeping the enolments intact in the db...but pointing the enrolments to the new payment if it succeeds. How do we know if the user aborts? Do we kill the enrolments on session timeout?
b. Fail the enrolments bringing them back to where they were before... i.e., they might be too late to enrol if places are filled. This is easy and in fact should be how it currently works by virtue of the enrolment controller living in the page rather than the session.

[2] Perhaps we do nothing, as if they adjust their shortlist they can simply choose to enrol the students already in their session in these.