-- Migration to Project WOnder 4.0.0.10 (+ patches) --
So, this involved ripping out or refactoring a number of things that were either no longer needed or that needed to function in a different way (e.g., entity munging on startup). Too much to mention.

Of note: (Basic guidelines for using WOnder)
1) extend ERXApplication
2) extend ERXSession
3) use ERXEC.newEditingContext()
4) Properties:
er.extensions.ERXApplication.useEditingContextUnlocker=true
er.extensions.ERXEC.defaultAutomaticLockUnlock=true
er.extensions.ERXEC.useSharedEditingContext=false
er.extensions.ERXEC.defaultCoalesceAutoLocks=true

With regard to the patches mentioned, I've submitted a couple of patches for things I was doing that needed support in WOnder. One of which has already been applied to head, the other is pending. 

When we're able to go for Java1.5 deployments then we can move to cvs head which will contain the patches and thus do away with my build.