Restoring replication credentials test behaviour.
The server is the only one that sets the credentials to false unless the client wants to do it if, for example, there's a blank username.

The client sets this flag to null to indicate that the server should test the credentials. If it fails the server will set it to false allowing a warning validation to be displayed in the client.