Fixing setting of backup preferences.
An exception was getting thrown as the formatter was set and then nullified at runtime when setting the custom component listener. TextFieldWithLabel now only sets the formatter explicitly if getFormatterForKey doesn't return null... so that if a formatter was already set it is not removed.

Also PreferenceViewController now returns the Integer formatter for those pref keys that are of an int type.