[squish 8769]
Ensuring contact's relationship to student is reset along with the isStudent flag (if the student should be discarded, because it was newly created by the controller but then thrown away).

This should ensure that students are not unintentionally committed with the context and (hopefully) maintain the relational integrity.

However, I would say that it may be better for the EnrolmentLineController to be the one that determines whether or not the isStudent flag is set on the contact... but that can be reviewed later. See the comments in QuickEnrollContactEditController#valueChanged.