[squish 8920] Fixing validation so that existing students can enrol.
This relaxes validateKEY property validation methods for normal validateForUpdate operations by having a specific flag (setStrictValidation) as false by default, but true for specific situations such as StudentAdd ... and later for login.

i.e., validateKEY methods are called during the validateForInsert/Update phases along with validateForSave -- rather than as thought for simply form values taken.

So with this new flag I can test in these methods whether or not to apply strict validation or not.