Fixing problem where batch count mismatch occurs.
Background:
Willow has a batch size of 10, fetches those 10 objects, attempts to translate each of them to SOAPEnterpriseObjects and ends up with less than 10 soap objects to send if one or more of the 10 fails to translate. This then causes an illegal state exception to be thrown back to angel when the next batch is attempted to be fetched due to not being able to release the current batch in willow.

We now simply increase the numberOfAttempts for any such records left over and so we can successfully release the current batch and move on.