Backtracking cache management.
Pages that implement SingletonCaching, when added to the pageCache, clear any previously held pages. i.e., you can't backtrack any further than these pages. Additionally SingletonCaching has a single method, restorePage, that allows the component to decide which component page should be restored. e.g., for LongResponse this returns the result of calling refresh().

The page cache stores transactions similar to WO's session transactions. The difference being that the transaction can hold either a component with its formValues etc OR a component type -- meaning that when restored a new page of that type is created in the current context.

Still need to think about what to do with the cached formValues. But this seems to be sufficient for testing.