[squish 9527] Updated stubs that include isCompany boolean flag for tutor.
Updated logic in translateToSoap/updateFromSoap to pass the lastName as companyName, firstName, and lastName (if flag is true) -- otherwise as is.

Backporting this to branch would be desirable to allow Companies to log into the Tutor Portal on the website.

Requires re-replicating all tutors after update to fix the data in willow.