squish[10868] action cog bug - becomes disabled after say cancelling an enrolment:
This bug not only happend in enrolment, it occured also in "classes" while "cancel" and "duplicate". and it affected not only the Cog button, also the "+" and "-" button was disabled. The reason was that after these actions a dialog appears which automatically disabled the buttons on the behind frame.

Furthermore I found out, that also after deleting a record in that outputlist these buttons were disabled (also a dialog appears)

Now a refresh is implemented.

all fixed