squish[10146] oncourse server shutdown - derby exception
fixed exception handling.

Derby documentation: 
A successful shutdown always results in an SQLException to indicate that Derby has shut down and that there is no other exception.

The XJ015 error (successful shutdown of the Derby engine) and the 08006 error (successful shutdown of a single database) are the only exceptions thrown by Derby that might indicate that an operation succeeded. All other exceptions indicate that an operation failed.