bug fixed for public synchronized void disposeControllers()
when terminating listcontrollers it is necessary that all their children/editcontroller are terminated first.

The error occured while logging out and show the login frame again. Sometimes a list (e.g. contact or course list) was still shown after logging out. In ControllerManager all generalControllers are saved in a map and there was no order how to dispose the controllers.