All the VET units of competency from a course are exported now (and exactly once)
Natalie wrote me:
"Instead of seeing the course name, as set by the user in the file, you should see a list of the units of competencies instead as the course name. For example, a course called 'Intro to computers' with two units attached, should show those two unit names and codes, but the same course without units attached should list in the NAT00060 as 'Intro to Computers'"

is working now, the algorithm already existed, but had a bug.

squish[11157]