SSL preferences added
In the general preference dialog we have "Enabled services" and "Features not enabled" where we list "Secure client server connection (SSL)". And if the user has paid we display it with a green tick and also display a checkbox where the administrator can tick "Allow only SSL connections".

the Preference table got 2 new entries:
-licence.sll: SSL connection is possible/is paid -> change it to true
-services.ssl.only: connection is only with SSL possible