Authentication against Microsoft Active Directory via LDAP implemented.
When a user tries to login on the onCourse client we have now the possibility to check if that user name and password exists in a LDAP server. If not the login will fail, if yes a normal login happens.

Like before the user will be created normally in the onCourse client and assigned to some acls. There is NO kind of synchronisation between onCourse client and the LDAP server, only a check. 

Parts of the functionality is working only for Microsoft Active Directory. I created the following entry on the Windows Server 2003 maschine:
cn: matthias moeser
dn: CN=matthias moeser,CN=Users,DC=winsrv03,DC=ish,DC=com,DC=au
sn: moeser
userPrincipalName: matthias@winsrv03.ish.com.au
userPassword: Halle-80*
(empty uid)

For authentication against that LDAP server the following steps are required:
1. Go to the “General Preferences” -> “Services” tab where we list “Enabled Services” and “Features not enabled”. If "Authentication against LDAP server" appears under "Enabled services" you can tick the checkbox "Authenticate every user against LDAP". After that you go to the tab "LDAP Authentication" and fill out all the fields:

Host: 203.29.62.133
Port: 389
Search base: CN=Users,DC=winsrv03,DC=ish,DC=com,DC=au
Attribute type for user name: sn
Attribute type for password: userPassword
Principal: matthias@winsrv03.ish.com.au
Password: Halle-80*

Choose "Simple Authentication" or "SASL Authentication + secure communication channel".

And now you need to create a new onCourse user with the name "moeser" and password "Halle-80*". In that test example it is the same user we use for security authentication and onCourse authentication.

The next time you log in the LDAP server verifies you.

The tricky part is to provide a secure connection if required. Generally there are two possibilities:
1. Using of SSL:
The implementation and the use is quite complicated: the server has to provide us a certificate which we load on the client. So we would have to implement a administration area for LDAP certificates.

2. SASL: Simple Authentication and Security Layer (RFC 2222)
LDAP and JNDI provide us with another way to encrypt and protect the authentication and also the communication channel. see

http://java.sun.com/products/jndi/tutorial/ldap/security/sasl.html
http://java.sun.com/products/jndi/tutorial/ldap/security/digest.html

For that approach we do not need to import any certificate. The implementation is server/vendor specific. I implemented and tested it against Microsoft Active Directory. (Choose in combobox "SASL Authentication + secure communication channel")

Furthermore in Microsoft Active Directory the user's password must be configured to be stored using reversible encryption: right click on user -> "Properties" -> "Account" -> tick: "Store password using reversible encryption". Also after the reversible encryption option has been set, the user must change their password for the password to be stored using this format: right click on user -> "Reset Password". See also:

http://www.forumeasy.com/forums/thread.jsp?tid=115170863235&fid=ldapprof5&highlight=Why+DIGEST-MD5+Authentication+Does+Work


Open issue:
we should think about if really every user must be authenticated against the LDAP server or if we have some independent admin. Because if someone typed in wrong LDAP parameter and the next time he logs in he will have a failed login. But at that time there is no way to correct the LDAP parameter because for that he needs to be logged in.