Backporting 6588:
All occurences of sql data type VARCHAR(32000) in Derby removed. Now CLOB(32000) is used instead.

The mapping in Cayenne Modeler was updated from VARCHAR to CLOB

This was necessary because in MSSQL server we got the exception
com.microsoft.sqlserver.jdbc.SQLServerException: "The text data type cannot be selected as DISTINCT because it is not comparable."
(so we had to switch the mapping in Cayenne from VARCHAR to CLOB)

This change affects a lot of the client, i tried to click through as much as possible.