Added ERXMigrations to WillowEOAccess and set it to run when WillowAdmin is launched.
When WillowAdmin runs, it will automatically create a "_dbupdater" table in the willow database, provided the willow user has table creation privileges. After that it will run migration.Willow0.java, which will create the NotificationTemplate table, and record the migration level as 0. This commit also includes two empty migrations Willow1 and Willow2 just to test the process. At the end of this process, the migration level will be 2 and further launches of WillowAdmin will not do any migration until migration.Willow3.java is added to the applciation.

The settings in the Properties file are under the heading # MIGRATIONS. Note that migration zero assumes that the database is the current live database - it does not create all the existing tables.