ClockingIT #50: HTML emails
onCourse sends now multipart emails out with a text/plain and text/html part. For that purpose 2 tabs in the send mail dialog are provided: "Plain" and "HTML". In the "HTML" tab there is a field for copying html code and another field for specifying a web url which content is downloaded. 

So 3 textfields are provided, minimum 1 must be used. 

Please remind that our database column can hold only up to 32000 character. That means only small websites can be used. For example http://www.sydneycommunitycollege.com.au/ has already 19240 characters.

When using a web url as source we insert a "base href" after the <head> tag so all links to css and images are working properly. But viewing it on a website of Gmail, Yahoo, Hotmail ... is a problem because the links to css are skipped (everything outside the <body> tags is ignored). See

http://xavierfrenette.com/articles/css-support-in-webmail/

Gmail:
* Keeps only the HTML between the <body> and </body> tags. Discards the rest.
* Removes any <style> element and its content, even if it's within the <body> element.
* Removes any <link> element.
* The font size is bigger in <table> elements, so you'll probably have to force it with: <table style="font-size:/*your size*/;"><!-- table content --></table>

Hotmail
* Keeps only the HTML between the <body> and </body> tags. Discards the rest.
* The <style> element and its content is preserved if it's inside the <body> element.
* Removes any <link> element.

Yahoo! Mail
* Keeps almost intact the complete original HTML sent and puts it directly in a <div>. So if you had any of these in the HTML, they will be found in the middle of the Yahoo! Mail HTML: DTD declaration, comments, <html>, <head>, <title>, <style> and <body> elements, etc. But, the DTD declaration won't be taken in consideration by the browser since it's in the middle of the page, same thing for the <title> element;
* Will rename the <body>, <meta> and <link> tags <xbody>, <xmeta> and <xlink>, so they will be "disabled". This means that no external stylesheet can be linked to the document through the <link> element.


Open issue: i want the field for the URL as a TextFieldWithLabel. But for some reason it works only as TextAreaWithLabel (public void setValueForKey(String key, Object value) is not called). I talked to Marcin, he  has a look on it.