SSL connection between client and server.
The client accepts any certificate send by the server and prompts the user a message where he has to confirm if he accepts the certificate which is saved.

The server is now that far that he creates a keystore with private public key and that stream can be stored in the database, see ish.oncourse.server.security.KeystoreGenerator

But at the moment it is not yet finished to load that and initialize the jetty http server, at the moment a fixed keystore is read. see at HttpServer 138.