Made new component WebDetailEditor that can be used in place of DetailPage to edit tags (web pages). This component includes an admin toolbar, which is placed at the top of the page by CSS. In normal mode, this component wraps a DetailPage, but when editing the displayed page, the web page editor is displayed. If no page is displayed, this component displays a site map that allows navigation to any page, including invisible pages.
Modified DetailPage to show WebDetailEditor instead of DetailPage if the Preference feature.editwebpage is true and the logged in WillowUser has web editing rights.

Created a TagList component that displays the tag (web page) hierarchy, with various modes allowing selection of tags and/or changing a tag's parent.

Created AdminLogin and AdminLoginWrapper components that allow a WillowUser to log in, separate from the student/tutor login.