Fix for vertical inheritance bug where two sub-entities are being inserted simultaneously into the db. NeededByEOF<index> was not being resolved to the correct attribute.
Patches have been submitted to both Apple and Wonder. The fix is in ISHEntity.anyAttributeNamed. However, this needs to be in com.webobjects.eoaccess for various superclass default or protected access level instance methods and fields.

Note: ISHModel is added to enable subclassing EOEntity as is ISHModelGroup for EOModel.
Other classes are there for adaptor logging and so forth.