Task 394:
In 70.sql (Derby) 

Table ConcessionType is dropped then recreated. The new table is created without a credentialExpiryDays column in Derby. 
Which is added in 71.sql

For MSSQL, MySQL and Oracle sql files, ConcessionType is created with credentialExpiryDays in 70.sql, which means that 71.sql is unnecessary for these dbs.

Proposed fix:
Add the column at table creation on 70.sql and remove 71.sql