Migrating to v3 of stubs. Note: payment-processing/enrolments-reservations still utilises v2 for now but in order for the test cases to compile when v2 is ripped out they've been adjusted for compiling (and don't work yet).
So replication should work as normal just with some new entities. 
Note also: because of the way replication works for contacts/students/tutors it is not necessary for a Student/Tutor to be created without creating a contact. Thus validation on the server has been relaxed further to not throw an exception if a contact doesn't exist for them.