It is now possible to choose any logo you wish for the printed reports.
Go to File -> Preferences -> General -> tab Printing. There you see all available reports and on the top of that a button "Set print header logo". Clicking on that opens a dialog which displays the current logo and where you can import any image (smaller that 2MB).

The image can have any height or width, it is adjusted to the correct size on the report.