New enrolment confirmation printout (see attachment)
- done like in the mock-up
- address field sits in the right position in the DL envelope
- html tags removed

one open issue:
in the session subreport details band there is an iteration over room.site.name, etc. But after that iteration we need the first value of room.site.name again. But we get only the last value from the iteration, not the first.

i tried to change the evaluation time, but no success.