Fixing handling of offline enrolments and payments or where procesing of either is disabled due to licencing.
If payment processing is disabled (e.g., non-credit card payment) this is succeeded whilst allowing the enrolments to proceed with processing against the gateway.
If enrolment processing is disabled (e.g., website not enabled) then these are succeeded whilst allowing the payment to proceed.
No longer throwing an exception if the payment is a non-credit card payment. This simply means handling for success and updating the date banked.
If the gateway is unavailable then the payment/enrolments are marked as queued (if processing is enabled for either).
Lots of graceful handling of the different paths to processing.