I moved the session details list behind the directions hints of the first session.
The reason:
In the session subreport we iterate in the details field over all sessions and print out their rooms and sites (e.g. room.site.name). If we print out after that a single room/site for the first session (in a footer field), we get instead the value of the last room/site.

I tried a lot so far like change the evaluation time (e.g. after evaluation group) but it did not work.