CIT #589 fixed.
1. when you tried to save a file and clicked into a subfolder, the FileChooser lost the file name and displayed only the folder which caused an exception after clicking "save"

The reason was that the FileChooser was able to select files and(!) folders. So by clicking into a subfolder this subfolder was selected and not the file itself any more.

It is now possible to set the FileSelectionMode of the FileChooser to FILES_ONLY and FILES_AND_DIRECTORIES via a paramter.

I reproduced and fixed this bug also for other places:
- save report as excel
- send message: MailingList.csv
- classList: export to xml


2. check if file is already existing and display dialog if user wants to override it (this is also for the excel part implemented)


3. catch and display message if user tries to override file which is used by another program (also for the excel part implemented)