CIT #542 Printed tax invoice, tax amount
fixed.

"tax amount" is a variable in jasperreports and is calculated by multipying the fields "quantity" and "taxEach" of the table InvoiceLine.

I reproduced the bug and found out that "tax amount" is always not printed when the values of "quantity" and "taxEach" did not change from one line to the next.

The reason was that in iReport in the common properties of according textfield the checkbox "print repeated values" was unticked.