CIT #613 Print header address block
done.

In the past we used the avetmiss fields in the General preferences for creating the headers on our reports.

Now we have a more flexible solution. Every college can create an own print header using a new textarea in the General preferences under the tab "print". 

On the report we then have on the top the college name (specified in the General preferences on the top of all tabs) and straight beneath the text of that new textarea.

With that approach it is possible for example for Tasmania to specify their head 
"Department of Education" \n
"Community Knowledge Network"

For colleges that already use onCourse I wrote a DataUpgrade to copy the values of the avetmiss fields to that new textarea.

The textarea allows you to enter maximum 6 lines with up to 82 characters per line to enforce that everything fits into the report title.