solved many problems in one go: - the CPU usage on QE - the drawing problems in many places
all was caused by the profile picture component, all went away with update of the mistral library (by the way there seem to be some development going on again on this project)
The bug was in the mistral library itself, paint method was causing invalidation of the component so the next time event dispatch thread was looking for 'dirty' regions it was painting the image again and again. about 10 times a second. 

Need now to test the other bits of gui.