[#410] Class by subject print error
A Nullpointer Exception happened in CourseClassPrintableObject.java line 160.

I cannot reproduce it, but it happens in that line:

logger.debug("Full Name = " + enrol.getStudent().getContact().getFullName() + " Sex = " + enrol.getStudent().getContact().getIsMale() +	" is Company = " + enrol.getStudent().getContact().getIsCompany());

That means that either enrol.getStudent() or enrol.getStudent().getContact() is null which should not(!) happen (wrong data in database?)

It belongs to a part where the number of male and female students is calculated (if it is specified).

I surround it with a try catch, because that part is only interesting if a sex is specified