added cayenne model and db schema for products.
I consider this a highly unstable, and if it will happen that the schema has to be adjusted I might just adjust the existing sql code and ask everyone to dump their data files. Maintaining safety in the trunk is too hard when we have few weeks to develop tons of new things.

#44