#748 XSLT based export
Step 1 implemented: infrastructure for exporting xml files.
If you open a list view you have now in the file Menu below "Print" an entry for "XSLT Export". At the moment we can export 2 xml files:
- classes
- contacts

What else is necessary? And what informations/columns must be contained?


In step 2 we have to apply xsl files to the xml files on the server. We probably need a new table "XSLTransformation" where we save xml and xsl files:

xml: to get the xml files to the server
xsl: to add and edit xsl files on the client

(analog to the table "Report")