#912 Creating sample database - asks to override when choosing desktop
done.

I also reverted the distinction for the file selector between mac (java.awt.FileDialog) and other systems (javax.swing.JFileChooser). If we start the server with no database we have to select a folder. Using FileDialog on a Mac is possible by setting a java system property "apple.awt.fileDialogForDirectories" to "true". But on Windows and Linux i did not found a way to select folders with java.awt.FileDialog. There we have to take javax.swing.JFileChooser.