Update miglayout to 3.6.3 bug fix release. Just some small bug fixes:
Added Linux/Gnome platform defaults. This means that Gnome Human Interface Guidelines' gaps and sizes are now correct on Linux. 	joeluckelman
fix 	MigLayout in a Serialized component throws a NPE. 	mikaelgrev
fix 	"skip" skipped cells before they where checked if there was a spanning component. This meant that "skip" did sometimes not skip a cell, instead it skipped the cell that would be skipped anyway since there was another component (spanning) there already. This is unlikely to affect many and if it does you are probably overusing the "skip" keyword anyway. :) 	mikaelgrev
fix 	"skip" did not take automatic layout wrapping (wrap in the Layout Constraints) into account. So if you skipped the last cells on a row the component did not end up in the next row as expected but on the same row. "skip" now alwats skips the correct number of cells and move to new rows as needed. 	mikaelgrev
fix 	If "newline size" or "wrap size" was used with "span" the size was attached to the wrong row. Also, if there was a wrap on one line and a newline on the next the extra row was removed, which was also a bug. 	mikaelgrev
fix 	Fixed a strange srinkage of some components due to calling Window.getPreferedSize() within the layout loop. This seem to be a workaround for a Swing bug rather than a big in MigLayout but it is hard to know.