[#998]
Removed the extra DD, DL and DT html elements from the source code that were conflicting with the new styles that are inherited from the base.css to display forms and form labels. 

The change of syntax is aimed to also reduce the amount of code in the component. 

Removed the extra CSS that was adding padding to the #sidebar that I had accidentally left when reworking the sidebar component.

[#103]

Similarly to the login and enrollment pages, I removed all of the extra DL, DT and DD elements to simplify the code and take advantage of the new CSS styles.