#289 Server config file
It is now possible to use a config file which has to be in the same folder as the application. When the server starts it checks if a file called "onCourse.cfg" exists and reads out its data.

if "port" is 0 then non-SSL connection are disabled and if "ssl_port" is 0 then SSL connection are disabled.

In the task is also written to specify ip addresses and log path. How and where do we do this in the server java code? Marcin?


onCourse.cfg should look like that:
# Specify the port as 0 to disable non-SSL client connections
port=8181

# Specify the port as 0 to disable SSL client connections
ssl_port=8182

#database, e.g. db=jdbc:mysql://delish.ish.com.au:3306/onCourse?user=matthias&password=topsecret
db=C:\onCourse\test2\onCourse.iocdata