# 519 fixing problem with expired sessions causing failing redirects on the site.
By default WOApplication, if refusing new sessions, returns a 301 redirect to /cgi-bin/WebObjects/AppName.woa.
Because we're using mod_rewrite and are redirecting when cgi-bin is there but shouldn't be the browser again gets redirected.
The fix is the override the public, yet undocumented, method app._newLocation and feeding the result into app._rewriteURL if it was a direct action based request.

This way, people doing their general browsing can continue to do so unhindered.