i feel like a killed a fly with a 5000MT nuke...
this is a patch allowing to untick a tutor in a session.
why so complex? cause I needed to display a non-cayenne object in lists, so I needed to employ a interface.

I understand that the fix is big, therefore I dont know if we are going to include it in the final release. On the other hand I have no other way of fixing the bug we had.

#1213