XSL files added for: - Contact csv file - CourseClass csv file - course csv file - enrolment csv file - invoice csv file - room csv file - site csv file and - Indesign brochure from class list view.
This needs to be tested when debugging of the XML export is finished (Lachlan got a list what I found). Furthermore at the moment the transformation only works if i remove 

xmlns="http://www.ish.com.au/schema/2009/onCourseData"

from the <data>-Tag in the XML files.

Also the Indesign xsl file is adapted together with Lachlans extension of the CourseClass export (adding site, room and tutor also to session and "average Hours" to class). There is only one issue open, it is with the <Tag> tags. The Indesign XML has to be like: 

<Tag>
<Title>Subjects</Title>
<Description>null</Description>
<Tag>
<Title>arts</Title>
<Description>null</Description>
<Tag>
<Title>Writing</Title>
<Description>null</Description>
<Course>
...
</Course>
</Tag>
</Tag>
</Tag>

I haven't understand yet the connection between course/class and <tag> in the CourseClassExport.xml. So that would be still open for Indesign, sorry.


Happy easter
Matthias