Modifying Windows service installer to have the following behaviour:
If no onCourse.cfg file is found, create onCourse.cfg and populate it with the data dir specified by the user installing - this is the only line in the file eg. db=WINDOWSPATH
If this file does exist assume that the contents are valid and skips the user entering the data directory part of the installer

In all cases the service is starts automatically.