#1328 report/printing related exception #1328 report/printing related exception
fixed.

The problem was the webdescription of the related course for an enrolment confirmation report. The text of the webdescription is transformed via ish.util.Textile to html. Doing that we had some bugs with lists (<ul>) and blank lines. This resulted in an incorrect html structure.

When printing a report we pass jasper such a html text and jasper detects it as html and parses it with Xerces SaxParser. Having an incorrect html causes the exception. Actually that doesn't matter because the report is still printed with no problem. But in the report we transform it to normal text and with incorrect html there might be some tags not be removed.

I tested a lot of enrolments. Should there be such an exception again a user can also try for the first to modify the course webdescription.