#1913 updating quicksearch (JSON) from Ari's spec - only show tags that match search term - show arbitrary filtered list of courses that match (picking max courses alphabetically sorted where courses are grouped by tag and those tags are descending sorted by how many courses they have each) - no longer showing course detail when less courses found - possible fix for JSON response by removing blank lines in response (more testing needed) - server-side limiting of items returned - removing obsolete json
must say - I'm not convinced on the usefulness of the courses section when more results are found. In that case I'd prefer to see their tags in the list and no courses shown. i.e., perhaps it'd be better to only show some courses when there's a small amount found.