#2239 fixing (I believe) info replication
When syslog is used - and not correctly configured - commonts-http-client seems to deadlock and so the info replication handler gets stuck. This also causes the server to fail to shutdown (on mac / unix).

I've added a log_default.properties which is a copy of the log_win.properties file and adjusted the code to refer to it instead.