Adding option of strict filtering to the visibles controller. The default is false (for Courses not being filtered by found classes).
However, when the /classes action is called strict filtering ought to be on so that the list of courses found is relevant to the browsed classes tag.

Similarly the CourseListItem needs to be aware of whether strict filtering is on to determine whether collapsing of non-matching classes is relevant and for obtaining the list of classes (either course.currentClasses for unstrict or controller.displayedClassesForCourse).

Essentially this ought to fix Juan's term dates class listings (which regressed from the previous version by simply listing classes found).