#2721 fixing QE so that it saves again.
A real cleanup is required in my opinion for how the controllers interact with each other. They're too easily broken. Case in point: there's two ways to save the context - saveRecord and commitChanges. Quick Enrol depends on the former being called whereas prior to this commit only commitChanges is called from the save button in the view. I don't understand why there's a need for two ways to save the context. There should be pre-post save hooks if needed, of course.