adding 404 request handler for bogus request safety.
Essentially when app.dispatchRequest is called it asks the request for its requestHandlerKey() which in turn internally calls request._uriDecomposed() which throws an NSForwardException if the uri is incapable of being decomposed. The assumption was that uri's are valid -- but our friendly spiders are a little more cunning.

So the new ISH404RequestHandler simply returns our 404 page for any such bogus requests.