fix fetching of classCosts for contact.
TODO: We need an isFinalised on ClassCost or this is going to increasingly get expensive. We can only use the start date of a class to filter costs that become relevant, but the end date can't be used. So currently every class that is either running or has run is in scope for fetching costs.