#15 fixing layout excess gaps (caused by custom components)
adding default implementation for CustomComponentInterface dispatchEvent(foo) which can be used by custom components not extending CustomComponent.

Adding cleanroom implementations of JCheckbox and JLabel which implement CustomComponentInterface. The advantage of these are that they are standard javax components with extensions for our events.

We can chat next year about necessary deprecations for Checkbox and LabelComponent which don't layout as nicely being jpanels.