Add convenience methods to Session to add and remove TutorAttendance;
Finally - fix selection of tutors for one or all sessions of a class. The problem was that the AttendanceController was creating new TutorAttendance records for every tutor in every session. The lack of a record is supposed to indicate that the tutor is not scheduled to be at that session, so creating new ones is not helpful. If in future we want a real attendance record for each tutor in each session, then we need a new flag in TutorAttendance that indicates they are not scheduled to attend.
[Task #23]