Because of the difficulty getting the Calendar custom component to work with the edit controllers, the following major refactoring was done:
Deprecate Calendar;
Add new packages for timetable controllers and components;
Add new CalendarView (ViewPanel) and CalendarViewController (ViewController, child of TimetableViewController);
Add a list of keys that trigger various redraw operations, and use setValueForKey() from the child controllers (as well as local changes) to trigger these;
Fix a number of the outstanding cross-component change-triggered display update bugs.

Note that there are still a number of outstanding display and functionality issues, but this version is better than the previously committed version.

[Task #23]