Backporting #2632 fixing find conditions for date/time.
The dates were being mangled from what the display formatter had already confirmed for the user. So, for example, 'after' turned the date into the following midnight even though the display formatter allowed the user to choose a time of day. The previous midnight in any case would have made more sense.

So I've made it work in this way:
- if they choose 'after' or 'before' there's no need to change the date from what was formatted.
- if they choose 'on' and the time of day is midnight then the intention is 'on' that day. If they specify a time of day, however, then if minutes is zero then they mean 'on' that hour, otherwise they mean 'on' the dateTime as entered.


Signed-off-by: Lachlan Deck <lachlan@ish.com.au>