#3355 (significant) refactoring of tagging structure, now utilising inheritance.
To all and sundry, fake node relationships are a thing of the past. Yay!

This is the first commit for this, and it thankfully compiles. But I expect that more adjustments will come in the coming days and weeks as bugs are further optimisations are found. I also plan on extending this work to Attachments (i.e., binary fake relations).

If you come across particular issues please let me know.


The background:
we need fixes now to the tag filtering problems which are really problematic of using fake relations. Then I realised that we didn't need inheritance on the Taggable entities, such as Course, as we'd previously thought but that we already had the database schema in place to make use of single-table inheritance for the join. This is cool, because (a) it doesn't require database refactoring which would cause all sorts of hassles moving forward, (b) we're not awaiting cayenne's vertical inheritance, (c) vertical inheritance would require painful refactorings of the database and replication ids and a straight-jacket for the entities anyway.

The implementation, or primary points to note:
- I've removed all Node* obj-entities
- Tag (replaces Node)
- TagRelation [abstract] maps all TagRelation attributes and the single relationship to 'tag'.
- concrete sub-entities are provided as joins for each taggable entity with a relevant qualifier on the entityIdentifier column. These are each called <EntityName>TagRelation and each provide a single toOne relationship called tagged<EntityName> to the relevant target entity.
- each Taggable entity adds a single toMany relationship to the relevant join <EntityName>TagRelation
- each Taggable entity now subclasses TaggablePersistentObject for the client and, in future, TaggableCayenneDataObject for the server.
- All common node-specific code that was previously in the common superclasses (e.g., PersistentObject) for all entities is now pushed down into these specific subclasses.
- TaggedPersistentObject, which is abstract, adds a couple of abstract methods such as taggedRelation which when implemented by subclasses returns the relevant tagged<EntityName> result making its usage quite generic.
- etc